chore: move addbutton and useControlledAccordion to common
